requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
psutil==5.8.0
APScheduler==3.7.0